+++ api/src/test/java/com/cantoalegre/api/ApiApplicationTests.java
@@ -1,10 +1,9 @@
 package com.cantoalegre.api;
 
+import com.cantoalegre.api.integration.AbstractIntegrationTest;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
-class ApiApplicationTests {
+class ApiApplicationTests extends AbstractIntegrationTest {
 
 	@Test
 	void contextLoads() {